Anindya office

+++ src/tasks/controller.py
@@ -1,13 +1,16 @@
-from fastapi import HTTPException
+from fastapi import HTTPException,status
 from src.tasks.dtos import TaskSchema
 from sqlalchemy.orm import Session
 from src.tasks.models import TaskModel
+from src.user.userModels import UserModel
 
-def create_task(data:TaskSchema, db:Session):
+def create_task(data:TaskSchema, db:Session, user:UserModel):
     data = data.model_dump()
     new_task = TaskModel(title= data["title"], 
                 description = data["description"], 
-                is_completed = data["is_completed"])
+                is_completed = data["is_completed"],
+                user_id = user.id
+                )
 
     db.add(new_task)
     db.commit()
@@ -17,8 +20,8 @@ def create_task(data:TaskSchema, db:Session):
 
 
 
-def get_tasks(db:Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db:Session, user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 
@@ -30,7 +33,7 @@ def get_one_task(task_id:int, db:Session):
     return one_task
 
 
-def update_task(data:TaskSchema, task_id:int, db:Session):
+def update_task(data:TaskSchema, task_id:int, db:Session, user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task Id is not found..")
@@ -38,6 +41,9 @@ def update_task(data:TaskSchema, task_id:int, db:Session):
     # one_task.title = data.title
     # one_task.description = data.description
     # one_task.is_completed = data.is_completed
+
+    if one_task.user_id != user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="You are not allowed to update this task")
 
     body = data.model_dump()
     for field, value in body.items():
@@ -50,10 +56,13 @@ def update_task(data:TaskSchema, task_id:int, db:Session):
     return one_task
 
 
-def delete_task(task_id:int, db:Session):
+def delete_task(task_id:int, db:Session, user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task id is incorrect..")
+
+    if one_task.user_id != user.id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="You are not allowed to delete this task")
 
     db.delete(one_task)
     db.commit()
